delete button only shows to the user

+++ src/react/components/DeleteMessageButton.js
@@ -1,30 +1,32 @@
 import React, { Component } from 'react';
 import { Button, Icon } from "semantic-ui-react"
-import { connect } from "react-redux"
+import { connect } from "../HOCs"
 import {deleteMessage} from "../../redux"
-
-// this.props.id
+ 
 class DeleteMessageButton extends Component {
     handleDelete = e => {
-        e.preventDefault();
-        this.props.deleteMessage(this.props.id)
-    }
-
+       const confirm = window.confirm("do you wnat to delete");
+       if(confirm){
+         this.props.deleteMessage(this.props.id)
+       }
+    };
     render() {
-        return (
-            <Button onClick={this.handleDelete} icon>
-                <Icon name='trash' />
-                Delete
-            </Button>
-        )
-    }
+      return (
+         
+        <Button icon onClick={this.handleDelete}>
+          <Icon name="trash" />
+          Delete
+        </Button>
+        
+      );
+  } 
 }
 
 export default connect(
-    state => ({
-        result: state.messages.deleteMessage.result,
-        loading: state.messages.deleteMessage.loading,
-        error: state.messages.deleteMessage.error
-    }),
-    { deleteMessage }
+  state => ({
+  result: state.messages.deleteMessage.result,
+  loading: state.messages.deleteMessage.loading,
+  error: state.messages.deleteMessage.error
+}),
+{ deleteMessage }
 )(DeleteMessageButton);
